chore: enable additional golangci-lint checks
Enable asciicheck, bidichk, durationcheck, errchkjson, errorlint, fatcontext, gocheckcompilerdirectives, loggercheck, makezero, nilerr, nilnesserr, nolintlint, reassign, spancheck, unconvert, and wastedassign.

Remove the legacy exclusion preset and fix the resulting findings in v2 and execution.

+++ v2/pkg/engine/resolve/resolve_test.go
@@ -6810,7 +6810,7 @@ func TestResolver_ResolveGraphQLSubscription(t *testing.T) {
 				messagesHeartbeat++
 			}
 		}
-		assert.Equal(t, int32(messagesToSendFromHook+messagesToSendFromOtherSources+messagesHeartbeat), int32(len(recorder.Messages())))
+		assert.Equal(t, messagesToSendFromHook+messagesToSendFromOtherSources+messagesHeartbeat, int32(len(recorder.Messages())))
 		assert.Equal(t, `{"data":{"counter":20000}}`, recorder.Messages()[0])
 	})
